PXB-3295 : Undo tablespaces are not tracked properly with lock-ddl=REDUCED
Problem:
--------
1.
Undo tablespaces are not tracked properly. Since undo tablespaces are not opened via
fil_open_for_xtrabackup(), they are not tracked as 'copied'. This leads to wrong
decisions in handle_ddl_operations.

2.
When new undo tablespaces are created,
Server doesn't write a MLOG_FILE_CREATE record and so these are missed by the tracking
system.

Fix:
----
1. track undo tablespaces that xtrabackup copies without lock (before lock state)
2. After lock is taken, undo tablespces are discovered again (after lock state)

With this before and after states, we now determine undo files to be deleted,
undo files to be copied. Truncated undo tablespace use different tablespace id, so
old undo file is marked as deleted and new version of undo tablespace is copied

For example undo_001.ibu of space_id 10 is truncated, the filename remains same
but it space_id becomes 11. xtrabackup creates 11.ibu.del for undo tablespace to be deleted.
then undo_001.ibu.new with space_id 11 is copied under lock